#40 refactor-feature/JD분석내용전체조회: 패키지명 오타수정
