+++ src/main/java/com/zb/jogakjogak/jobDescription/repository/JDRepository.java
@@ -1,4 +1,4 @@
-package com.zb.jogakjogak.jobDescription.repsitory;
+package com.zb.jogakjogak.jobDescription.repository;
 
 import com.zb.jogakjogak.jobDescription.entity.JD;
 import org.springframework.data.jpa.repository.JpaRepository;
