+++ src/main/java/com/zb/jogakjogak/jobDescription/repository/ToDoListRepository.java
@@ -1,4 +1,4 @@
-package com.zb.jogakjogak.jobDescription.repsitory;
+package com.zb.jogakjogak.jobDescription.repository;
 
 import com.zb.jogakjogak.jobDescription.entity.JD;
 import com.zb.jogakjogak.jobDescription.entity.ToDoList;
